login/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
Flask-SQLAlchemy
flask-jwt-extended
passlib
Flask-WTF
Flask-Session
